upadte

+++ frontend/src/components/Home/ViewTask.jsx
@@ -2,8 +2,10 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import Navbar1 from "./Navbar1";
 import "./viewtask.css"; 
+import { useNavigate } from "react-router-dom";
 
 const ViewTask = () => {
+  const navigate = useNavigate();
   const [tasks, setTasks] = useState([]);
 
   useEffect(() => {
@@ -24,6 +26,7 @@ const ViewTask = () => {
           Authorization: `Bearer ${token}`,
         },
       });
+      console.log(response.data.title);
       console.log(response.data.id);
       setTasks(response.data);
 
@@ -52,27 +55,36 @@ const ViewTask = () => {
     }
   };
 
-  const handleEdit = async() => {
-
+  const handleEdit = async(taskId) => {
+    if(!taskId) {
+      console.error("Error: taskId is undefined!");
+      return;
+    }
+    navigate(`/edit-task/${taskId}`);
   }
 
   const handleDelete = async (taskId) => {
-    try {
-      const token = sessionStorage.getItem("access");
-      if (!taskId) {
-        console.error("taskId is undefined!");
-        return;
+    const confirmdel= window.confirm("Are you sure want to delete this task ?")
+    if(confirmdel){
+      try {
+    
+        const token = sessionStorage.getItem("access");
+        if (!taskId) {
+          console.error("taskId is undefined!");
+          return;
+        }
+    
+        await axios.delete(`http://127.0.0.1:8000/tasks/delete/${taskId}/`, {
+          headers: { Authorization: `Bearer ${token}` }
+        });
+    
+        setTasks(tasks.filter(task => task.id !== taskId));  // Remove deleted task from UI
+        console.log("Task deleted successfully!");
+      } catch (error) {
+        console.error("Error deleting task:", error);
       }
-  
-      await axios.delete(`http://127.0.0.1:8000/tasks/delete/${taskId}/`, {
-        headers: { Authorization: `Bearer ${token}` }
-      });
-  
-      setTasks(tasks.filter(task => task.id !== taskId));  // Remove deleted task from UI
-      console.log("Task deleted successfully!");
-    } catch (error) {
-      console.error("Error deleting task:", error);
     }
+    
   };
   
 
@@ -99,7 +111,7 @@ const ViewTask = () => {
                     <td>{task.description}</td>
                     <td>{task.user}</td>
                     <td>
-                      <button className="task-edit-btn" onClick={() =>handleEdit()}>Edit</button>
+                      <button className="task-edit-btn" onClick={() =>handleEdit(task.id)}>Edit</button>
                       <button className="task-delete-btn" onClick={() =>handleDelete(task.id)}>Delete</button>
                     </td>
                   </tr>
